More appropriate watchdog timeout default message.
Was pretty worried when I saw this @me this afternoon while away from home thinking some sort of temperature warning threshold were met, or any kind of alarming scenario. Instead it was just a plugin on the Survival instance acting up and making the server's gameticks take a while, activating this watchdog message.

Changed to a more appropriate default message.